[gn] Make Android build work in standalone
This makes it possible to build for Android in Dawn standalone.
Previously it could only be done inside a Chromium checkout.

NOTE: No effort was made to see if the resulting binaries actually work,
only to make sure they're buildable.

Much of this configuration was copied from ANGLE, which is similar to
Dawn and already has a standalone Android build.

Note the revisions of various packages are not exactly the ones that
roll_chromium_deps.py will roll to, because those are hard to figure out
since chromium/src has moved past our DEPS version. They'll get in sync
with the next run.

Bug: 520153663
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/315515
Commit-Queue: Kai Ninomiya <kainino@chromium.org>

+++ src/dawn/tests/white_box/SharedTextureMemoryYCbCrAndroidTests.cpp
@@ -625,7 +625,9 @@ TEST_P(SharedTextureMemoryVulkanYCbCrParamsTests, SampleY8Cb8Cr8AHB) {
         auto ycbcr = math::Vec4f(cr / 255.0, y / 255.0, cb / 255.0, 1.0);
         auto rgb = math::Mul(ycbcrToRgb, ycbcr);
 
-        auto expected = utils::RGBA8(rgb[0] * 255, rgb[1] * 255, rgb[2] * 255, 255);
+        auto expected =
+            utils::RGBA8(static_cast<uint8_t>(rgb[0] * 255), static_cast<uint8_t>(rgb[1] * 255),
+                         static_cast<uint8_t>(rgb[2] * 255), 255);
         auto bottom = utils::RGBA8(expected.r - 1, expected.g - 1, expected.b - 1, 255);
         auto top = utils::RGBA8(expected.r + 1, expected.g + 1, expected.b + 1, 255);
